Feature/skale 1516 finalizing fast snapshot distribution system

+++ skaled/main.cpp
@@ -370,8 +370,8 @@ int main( int argc, char** argv ) try {
     // skale - snapshot download command
     addClientOption( "download-snapshot", po::value< string >()->value_name( "<url>" ),
         "Download snapshot from other skaled node specified by web3/json-rpc url" );
-    addClientOption( "download-target", po::value< string >()->value_name( "<port>" ),
-        "Path of file to save downloaded snapshot to" );
+    // addClientOption( "download-target", po::value< string >()->value_name( "<port>" ),
+    //    "Path of file to save downloaded snapshot to" );
 
     LoggingOptions loggingOptions;
     po::options_description loggingProgramOptions(
@@ -982,35 +982,20 @@ int main( int argc, char** argv ) try {
                               "prices_" + chainParams.nodeInfo.id.str() + ".db"} ) );
 
     if ( vm.count( "download-snapshot" ) ) {
+        int nImportResult = 0;
+        fs::path saveTo;
         try {
-            fs::path saveTo;
-
-            if ( vm.count( "download-target" ) )
-                saveTo = fs::path( vm["download-target"].as< string >() );
-            else {
-                string filename_string = "/tmp/skaled_snapshot_download_XXXXXX";
-                char buf[filename_string.length() + 1];
-                strcpy( buf, filename_string.c_str() );
-                int fd = mkstemp( buf );
-                if ( fd < 0 ) {
-                    throw std::runtime_error( "Failed to open " + filename_string );
-                }
-                close( fd );
-                saveTo = buf;
-            }
-
             std::string strURLWeb3 = vm["download-snapshot"].as< string >();
 
             std::cout << cc::normal( "Will download snapshot from " ) << cc::u( strURLWeb3 )
-                      << cc::normal( " to " ) << cc::info( saveTo.native() ) << std::endl;
-            bool isBinaryDownload = true;
+                      << std::endl;
 
             unsigned block_number;
-            {
+            {  // block
                 skutils::rest::client cli;
                 if ( !cli.open( strURLWeb3 ) ) {
-                    // std::cout << cc::fatal( "REST failed to connect to server" ) << "\n";
-                    return -1;
+                    nImportResult = 2;
+                    throw std::runtime_error( "REST failed to connect to server" );
                 }
 
                 nlohmann::json joIn = nlohmann::json::object();
@@ -1019,35 +1004,67 @@ int main( int argc, char** argv ) try {
                 joIn["params"] = nlohmann::json::object();
                 skutils::rest::data_t d = cli.call( joIn );
                 if ( d.empty() ) {
-                    std::cout << cc::fatal( "Snapshot download failed - cannot get bockNumber" )
-                              << "\n";
-                    return -1;
+                    nImportResult = 3;
+                    throw std::runtime_error( "download failed, cannot get bockNumber" );
                 }
+                nlohmann::json joAnswer = nlohmann::json::parse( d.s_ );
+                // std::cout << cc::normal( "Got answer " ) << cc::j( joAnswer ) << std::endl;
                 // TODO catch?
-                block_number = dev::eth::jsToBlockNumber(
-                    nlohmann::json::parse( d.s_ )["result"].get< string >() );
+                block_number = dev::eth::jsToBlockNumber( joAnswer["result"].get< string >() );
                 block_number -= block_number % chainParams.nodeInfo.snapshotInterval;
-            }
+            }  // block
 
-            bool bOK = dev::rpc::snapshot::download( strURLWeb3, block_number, saveTo,
-                [&]( size_t idxChunck, size_t cntChunks ) -> bool {
-                    std::cout << cc::normal( "... download progress ... " )
-                              << cc::num10( uint64_t( idxChunck ) ) << cc::normal( " of " )
-                              << cc::num10( uint64_t( cntChunks ) ) << "\r";
-                    return true;  // continue download
-                },
-                isBinaryDownload );
-            std::cout << "                                                  \r";  // clear progress
-                                                                                  // line
-            if ( !bOK ) {
-                std::cout << cc::fatal( "Snapshot download failed" ) << "\n";
-                return -1;
+            try {
+                bool isBinaryDownload = true;
+                std::string strErrorDescription;
+                saveTo = snapshotManager->getDiffPath( 0, block_number );
+                bool bOK = dev::rpc::snapshot::download( strURLWeb3, block_number, saveTo,
+                    [&]( size_t idxChunck, size_t cntChunks ) -> bool {
+                        std::cout << cc::normal( "... download progress ... " )
+                                  << cc::num10( uint64_t( idxChunck ) ) << cc::normal( " of " )
+                                  << cc::num10( uint64_t( cntChunks ) ) << "\r";
+                        return true;  // continue download
+                    },
+                    isBinaryDownload, chainParams.nodeInfo.snapshotInterval, &strErrorDescription );
+                std::cout << "                                                  \r";  // clear
+                                                                                      // progress
+                                                                                      // line
+                if ( !bOK ) {
+                    nImportResult = 4;
+                    if ( strErrorDescription.empty() )
+                        strErrorDescription = "download failed, connection problem during download";
+                    throw std::runtime_error( strErrorDescription );
+                }
+            } catch ( const std::exception& ex ) {
+                nImportResult = 5;
+                std::cerr << cc::fatal( "FATAL:" )
+                          << cc::error( " Exception while downloading snapshot: " )
+                          << cc::warn( ex.what() ) << "\n";
+                throw;
+            } catch ( ... ) {
+                nImportResult = 6;
+                std::cerr << cc::fatal( "FATAL:" )
+                          << cc::error( " Exception while downloading snapshot: " )
+                          << cc::warn( "Unknown exception" ) << "\n";
+                throw;
             }
             std::cout << cc::success( "Snapshot download success for block " )
                       << cc::u( to_string( block_number ) ) << std::endl;
-
-            snapshotManager->importDiff( block_number, saveTo );
-            fs::remove( saveTo );
+            try {
+                snapshotManager->importDiff( 0, block_number );
+            } catch ( const std::exception& ex ) {
+                nImportResult = 7;
+                std::cerr << cc::fatal( "FATAL:" )
+                          << cc::error( " Exception while importing downloaded snapshot: " )
+                          << cc::warn( ex.what() ) << "\n";
+                throw;
+            } catch ( ... ) {
+                nImportResult = 8;
+                std::cerr << cc::fatal( "FATAL:" )
+                          << cc::error( " Exception while importing downloaded snapshot: " )
+                          << cc::warn( "Unknown exception" ) << "\n";
+                throw;
+            }
 
             // HACK refactor this piece of code!
             fs::path price_db_path;
@@ -1068,17 +1085,25 @@ int main( int argc, char** argv ) try {
             //// HACK END ////
 
             snapshotManager->restoreSnapshot( block_number );
+            std::cout << cc::success( "Snapshot restore success for block " )
+                      << cc::u( to_string( block_number ) ) << std::endl;
 
-            fs::remove( saveTo );
         } catch ( const std::exception& ex ) {
+            if ( nImportResult == 0 )
+                nImportResult = 1000;
             std::cerr << cc::fatal( "FATAL:" )
-                      << cc::error( " Exception while downloading snapshot: " )
+                      << cc::error( " Exception while processing downloaded snapshot: " )
                       << cc::warn( ex.what() ) << "\n";
         } catch ( ... ) {
+            if ( nImportResult == 0 )
+                nImportResult = 1001;
             std::cerr << cc::fatal( "FATAL:" )
-                      << cc::error( " Exception while downloading snapshot: " )
+                      << cc::error( " Exception while processing downloaded snapshot: " )
                       << cc::warn( "Unknown exception" ) << "\n";
         }
+        if ( !saveTo.empty() )
+            fs::remove( saveTo );
+        return nImportResult;
     }
 
     // it was needed for snapshot downloading
